backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==0.28.1
langchain==0.0.340
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0